Textual Analysis and Machine Learning for CyberstalkingDetection
Cyberstalking and harassmenthave been identified as a growing anti-social problem. Be-sides detecting cyberstalking and harassment, there is the need to gather digital evidence, often by the victim. 
It's a problem for my HSP community who requires solitude to rejuvenate. Our community is very susceptible to these aggressive cyberstalkers leading up to offline stalking.