New  GRYPHON - Advance Security router
Made a police report on the cyberstalker/stalker sped up my shipment so he can tamper with it.  
On Amazon app, it said someone has received the package for me. 
So, when I want to return the item on Amazon, this is the screenshot of the return date, 2021.
And, he changed the "return date to Jan 31, 2021".
So he is watching my every move and making sure he has control of my electronic devices.